// src/type/sort.type.ts
export enum SortType {
  DEFAULT_SORTING = "Default sorting",
  SORT_BY_POPULARITY = "Sort by popularity",
  SORT_BY_AVERAGE_RATING = "Sort by average rating",
  SORT_BY_LATEST = "Sort by latest",
  SORT_BY_PRICE_LOW_TO_HIGH = "Sort by price: low to high",
  SORT_BY_PRICE_HIGH_TO_LOW = "Sort by price: high to low",
}